Separated quick items

+++ lib/widgets/quick_items/quick_items_widget.dart
@@ -3,14 +3,15 @@ import 'dart:async';
 
 // Flutter imports:
 import 'package:animations/animations.dart';
-import 'package:dio/dio.dart';
 import 'package:flutter/material.dart';
 
 // Package imports:
 import 'package:flutter_inappwebview/flutter_inappwebview.dart';
 import 'package:material_design_icons_flutter/material_design_icons_flutter.dart';
 import 'package:provider/provider.dart';
+import 'package:torn_pda/models/quick_item_model.dart';
 import 'package:torn_pda/pages/quick_items/quick_items_options.dart';
+import 'package:torn_pda/providers/quick_items_faction_provider.dart';
 import 'package:webview_flutter/webview_flutter.dart';
 
 // Project imports:
@@ -18,13 +19,11 @@ import 'package:torn_pda/providers/quick_items_provider.dart';
 import 'package:torn_pda/utils/js_snippets.dart';
 
 class QuickItemsWidget extends StatefulWidget {
-  final String webviewType;
   final InAppWebViewController inAppWebViewController;
   final WebViewController webViewController;
   final bool faction;
 
   QuickItemsWidget({
-    @required this.webviewType,
     @required this.faction,
     this.inAppWebViewController,
     this.webViewController,
@@ -36,6 +35,7 @@ class QuickItemsWidget extends StatefulWidget {
 
 class _QuickItemsWidgetState extends State<QuickItemsWidget> {
   QuickItemsProvider _itemsProvider;
+  QuickItemsProviderFaction _itemsProviderFaction;
 
   Timer _inventoryRefreshTimer;
 
@@ -54,6 +54,7 @@ class _QuickItemsWidgetState extends State<QuickItemsWidget> {
   @override
   Widget build(BuildContext context) {
     _itemsProvider = context.watch<QuickItemsProvider>();
+    _itemsProviderFaction = context.watch<QuickItemsProviderFaction>();
     return Padding(
       padding: const EdgeInsets.symmetric(horizontal: 5),
       child: Row(
@@ -85,10 +86,14 @@ class _QuickItemsWidgetState extends State<QuickItemsWidget> {
   List<Widget> _itemButtons() {
     var myList = <Widget>[];
 
-    for (var item in _itemsProvider.activeQuickItems) {
-      if (item.isLoadout && widget.faction) continue;
-      if (item.isPoints && !widget.faction) continue;
+    List<QuickItem> itemList = <QuickItem>[];
+    if (widget.faction) {
+      itemList = List.from(_itemsProviderFaction.activeQuickItemsFaction);
+    } else {
+      itemList = List.from(_itemsProvider.activeQuickItems);
+    }
 
+    for (var item in itemList) {
       Color qtyColor;
       if (item.inventory == 0) {
         qtyColor = Colors.orange[300];
@@ -98,7 +103,7 @@ class _QuickItemsWidgetState extends State<QuickItemsWidget> {
 
       double qtyFontSize = 12;
       var itemQty = item.inventory.toString();
-      if (!item.isLoadout) {
+      if (!item.isLoadout && !widget.faction) {
         if (item.inventory > 999 && item.inventory < 100000) {
           itemQty = "${(item.inventory / 1000).truncate().toStringAsFixed(0)}K";
         } else if (item.inventory >= 100000) {
@@ -166,21 +171,14 @@ class _QuickItemsWidgetState extends State<QuickItemsWidget> {
                           ),
             onPressed: () async {
               if (item.isLoadout) {
-                if (widget.webviewType == "attacks") {
-                  var js = changeLoadOutJS(item: item.name.split(" ")[1], attackWebview: true);
-                  widget.webViewController.runJavascript(js);
-                } else {
-                  var js = changeLoadOutJS(item: item.name.split(" ")[1], attackWebview: false);
-                  await widget.inAppWebViewController.evaluateJavascript(source: js);
-                }
+                var js = changeLoadOutJS(item: item.name.split(" ")[1], attackWebview: false);
+                await widget.inAppWebViewController.evaluateJavascript(source: js);
               } else {
                 var js = quickItemsJS(item: item.number.toString(), faction: widget.faction, refill: item.isPoints);
-                if (widget.webviewType == "attacks") {
-                  widget.webViewController.runJavascript(js);
-                } else {
-                  await widget.inAppWebViewController.evaluateJavascript(source: js);
+                await widget.inAppWebViewController.evaluateJavascript(source: js);
+                if (!widget.faction) {
+                  _itemsProvider.decreaseInventory(item);
                 }
-                _itemsProvider.decreaseInventory(item);
               }
             },
           ),
@@ -196,7 +194,9 @@ class _QuickItemsWidgetState extends State<QuickItemsWidget> {
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
               Text(
-                "Configure your preferred quick items",
+                widget.faction
+                    ? "Configure your faction's armoury quick items"
+                    : "Configure your preferred quick items",
                 style: TextStyle(
                   color: Colors.orangeAccent,
                   fontSize: 12,
@@ -239,7 +239,9 @@ class _QuickItemsWidgetState extends State<QuickItemsWidget> {
   }
 
   _refreshInventory() {
-    _itemsProvider.updateInventoryQuantities(fullUpdate: false);
+    if (!widget.faction) {
+      _itemsProvider.updateInventoryQuantities(fullUpdate: false);
+    }
   }
 
   Widget _settingsIcon() {
@@ -249,7 +251,9 @@ class _QuickItemsWidgetState extends State<QuickItemsWidget> {
         transitionDuration: Duration(milliseconds: 500),
         transitionType: ContainerTransitionType.fadeThrough,
         openBuilder: (BuildContext context, VoidCallback _) {
-          return QuickItemsOptions();
+          return QuickItemsOptions(
+            isFaction: widget.faction,
+          );
         },
         closedElevation: 0,
         closedShape: const RoundedRectangleBorder(
